Further simplified StorageConfig and moved PidManager
